requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
pytest
httpx

